Added MRTest +Minor change to MR
The MR test as it is right now doesn't seem to work right. It will always throw a error when you try to write to a MR. I am not sure if the error is in MR or in the way I have set up my tests.

I also changed it so that MRs always start with the status of free to avoid the problem of all MRs showing as FREE even when they are not.