Finalised setUp/radarContact Test
So I have moved the stuff for radarContact into setUp due to not being able pick and choose the order in which tests happen. This avoids the problem of things happening in the wrong order.

JUST REMEMBER THAT YOU CANNOT KNOW THE ORDER TESTS WILL HAPPEN IN SO DON'T SET UP SOMETHING IN ONE TEST THAT NEEDS SOMETHING FROM ANOTHER TEST TO HAVE HAPPENED BEFORE IT.