fixed UI in RTTest
Overthought the whole thing, running RTTest now creates the UI Element but it will throw a bunch of errors if you click on them. Need to work on that.

Also RTTest currently is not set up correctly, it basically functions like main() right now. I will change that down the road.